Created AdminInterface class and adminTest.py
AdminInterface class is in new AdminInterface.py file, this is included in adminTest.py thru import statement

Made adminTest.py to experiment with code and comment out RFID initialization for now while we work on just Admin related things.  This is to avoid integration related issues when we eventually bring this over to main branch.