Regex fixed to include all 3 possible combos of asset # formats
Will accept numbers that begin w/ E, e or 4.

If E or e, then will accept 7 digits afterwards

If 4, will only accept 6 digits afterwards

Search_UI_Message_Prompt keeps getting overwritten when an invalid number is entered, saying it can't be found rather than it is invalid.  This might be okay because it's letting the user know the input can't be retrieved and user will have an option to review their input for validity