Appropriate prompts for every search combination filter implemented
We should still go thru and try to break this again, maybe have a procedure to follow for testing the prompts.  I think this is fairly robust but we could have missed something, the code is enormous and I have not looked at some of these functions in a while.