find_files function now looks for all lettered drives including C:
Gives user option to look on C: drive with a warning that it takes some time to search thru.  Returns false if user selects no or if the file is not found on the C: drive either.

This also works if the file is placed in desktop, video, etc as those are all locations on the C: drive